requirements.txt
elevenlabs==1.3.0
openai==1.12.0
SpeechRecognition==3.10.1
pyaudio==0.2.14
python-dotenv==1.0.1
pydub==0.25.1
